apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pyyaml
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
schedule>=1.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
